Update modify_1week_5sec-interval.json
Adjusted values to represent the max number of stats events that would be calculated very 5 seconds in a 24 hour period. Sent the approximate files size for a larger  this should be lower, and set ti to keep 14 weeks of requests.  The files should be smaller in most envs so that Admins know the approximated max file use   For busy servers with long junction names this may change