Update modify_2week_5sec-interval.json
Adjusted values to represent the max number of stats events that would be calculated very 5 seconds in a 24 hour period. Sent the max approximate files size for a larger value actually in most uses cases should be smaller, I set it to rollover every day and keep 14 days of records. The files should be smaller in most cases I set it this way so that admins know the approximated max file space that will be used.  For busy servers with long junction names this may change.   If a server is restarted this logging will be turned off.